Moved ng templates to jscodegen
