+++ internal/jscodegen/ng-entity.tmpl
