+++ internal/jscodegen/ng-service.tmpl
