+++ internal/jscodegen/ng-standalone-entity.tmpl
